Phase 3 v1-fin-C: CPU tap-0 mutability + relax tap0_delay to float
Closes the last v1 read-path gap. CPU backend now rebuilds an
effective tap[0] from per-link `live` each slot before calling
apply_tdl_step{_fading}; CUDA already has equivalent host-side
refresh from v1-fin-B. tap0_delay_samples relaxed from int32 →
float so YAMLs with fractional delays (TR 38.901 TDL-A τ=2.5 etc.)
round-trip lossless through the control plane — uncovered by an
existing fractional-delay processing test that failed when the
old populate path rounded the YAML to an integer.

Surface changes:

- include/ocudu_gpu_channel/mutable_params.h:
    tap0_delay_samples: int32_t → float (struct still 32 bytes,
    cudaMemcpy contract preserved).
- src/mutable_params.cpp populate_mutable_params_from_yaml:
    drop std::lround; just static_cast<float>.
- src/control_server.cpp ParamSpec for tap0_delay_samples:
    is_int = true → is_int = false. Range [0, 1023] unchanged.
- src/cpu_backend.cpp apply_chain_to_link Tdl branch:
    build effective_taps + effective_polyphase from step.taps +
    state.live each slot. effective_polyphase[0] is rebuilt via
    compute_windowed_sinc_taps from the fractional part of the
    live delay — same helper apply_tdl_step uses at prepare, so
    CPU↔CUDA parity holds. Per-slot copy cost is O(n_taps) ≤ 32 —
    negligible vs the per-sample convolution.
- src/device_channel.cu refresh_tap0_from_live:
    use compute_windowed_sinc_taps for the fractional polyphase
    rebuild instead of the unit-impulse shortcut. Same parity
    fix as the CPU side.

Tests:

- tests/test_runtime_update_parity.cpp: new v1-fin-C section
  drives tap0_gain_db = 0 → +6 → -6 dB, asserts CPU output
  amplitude scales as 1.0 → ~2.0 → ~0.5 within 1e-3.
- tests/test_mutable_params.cpp: Case 1 expects 5.5F round-trip,
  not lround(5.5) = 6. Cases 2/3 zero-check switched to float.
- tests/test_control_server.cpp: Cases 4/5 inverted — fractional
  tap0_delay was rejected (int-only), now accepted; integer still
  works. Case 12 + 13 counter expectations bumped (5 applied / 6
  rejected from 4/7).

All 7 ctest cases pass.

Phase 3 v1 read-path rewires are now complete. All 7 v1 params
(path_loss_db, awgn_snr_db, cfo_hz, tap0_delay_samples,
tap0_gain_db, tap0_phase_rad, los_k_db) flow through `live` on
both backends; the snap-from-shadow mechanism updates them at
slot boundaries; the parity test exercises path_loss, awgn_snr,
and tap0_gain.

+++ include/ocudu_gpu_channel/mutable_params.h
@@ -35,8 +35,11 @@ struct MutableParams {
 
   // Per-edge channel params (consumed by apply_channel_kernel on CUDA or
   // apply_tdl_step{_fading} on CPU). Only the leading tap is mutable in v1;
-  // multi-tap profile swaps are v2.
-  int32_t  tap0_delay_samples  = 0;
+  // multi-tap profile swaps are v2. tap0_delay_samples is FLOAT so YAMLs
+  // with fractional delays (e.g. TR 38.901 TDL-A τ=2.5 samples) round-trip
+  // through the runtime control plane lossless; both backends recompute
+  // the per-tap polyphase coefficients from the fractional part on snap.
+  float    tap0_delay_samples  = 0.0F;
   float    tap0_gain_db        = 0.0F;
   float    tap0_phase_rad      = 0.0F;
   float    los_k_db            = 0.0F;
